refactor(modules): move templates to modules and add server/site tasks
- Move shell templates from taskrunner to server and site modules
- Add job handlers for server operations (provision, PHP, database, SSH, firewall)
- Add task definitions for MySQL, PostgreSQL, opcache, and services
- Add site deployment and rollback tasks
- Fix uint/uint64 type consistency in backup module
- Fix JSON handling for webhook provider data
- Update template engine to support module-level templates
- Remove unused team module scaffold

+++ internal/modules/backup/dto/requests.go
@@ -59,7 +59,7 @@ type CreateStorageProviderRequest struct {
 
 // UpdateStorageProviderRequest represents a request to update a storage provider
 type UpdateStorageProviderRequest struct {
-	ID       uint   `json:"id" validate:"required"`
+	ID       uint64 `json:"id" validate:"required"`
 	Label    string `json:"label" validate:"required,max=255"`
 	Provider string `json:"provider" validate:"required,oneof=s3 dropbox"`
